Compose 1.2.0 + Kotlin 1.7.20 + updates according to sniffer changes + small UI evolutions

+++ src/main/kotlin/fr/lewon/dofus/bot/handlers/characteristics/CharacterStatsListEventHandler.kt
@@ -2,7 +2,7 @@ package fr.lewon.dofus.bot.handlers.characteristics
 
 import fr.lewon.dofus.bot.gui2.main.scripts.characters.edit.global.CharacterGlobalInformationUIUtil
 import fr.lewon.dofus.bot.sniffer.DofusConnection
-import fr.lewon.dofus.bot.sniffer.model.messages.fight.CharacterStatsListMessage
+import fr.lewon.dofus.bot.sniffer.model.messages.game.character.stats.CharacterStatsListMessage
 import fr.lewon.dofus.bot.sniffer.store.IEventHandler
 import fr.lewon.dofus.bot.util.network.GameSnifferUtil
 
@@ -13,7 +13,10 @@ object CharacterStatsListEventHandler : IEventHandler<CharacterStatsListMessage>
         val characteristics = socketResult.stats.characteristics
         gameInfo.playerBaseCharacteristics = characteristics.associateBy { it.characteristicId }
         gameInfo.updatePlayerFighter()
-        CharacterGlobalInformationUIUtil.updateCharacterKamas(gameInfo.character.name, socketResult.stats.kamas)
+        CharacterGlobalInformationUIUtil.updateCharacterKamas(
+            gameInfo.character.name,
+            socketResult.stats.kamas.toLong()
+        )
     }
 
 }
